Auto regeneration of Scanner/Parser fixed.
The pregenerated versions of Scanner/Parser are now more isolated from the the
rest of the build source and are only used if they are really needed (in the
case of the lack of bison or flex).

Bison and flex versions are now checked to meet a minimum requirement.  If this
requirement is not met, the respective pregenerated code is used.

FlexLexer.h is now copied into the pregen-scanner directory when make is run.
This way, we can be sure to have the FlexLexer.h code that goes with the scanner
created by a particular flex version.