Fix yet one more initialization order problem (exhibited w/ MSVC)
Also changed static instances of Driver and registry to use stack allocations
and not heap allocations.  This way, the destructor will be called
automatically (I suppose I could have used a shared_ptr instead, but the impact
on the stack is very little--most of their data members lie on the heap anyway).